Move hooks to proper directory and reduce rate delta
